Nested all files in moodle::api module so the name moodle-api could be used in ruby gems. Unfortunately moodle was taken.
