+++ README.md
@@ -26,17 +26,17 @@ Different ways to configure gem.
 
 ```ruby
 # Pass block to configure
- Moodle.configure do|c|
+ Moodle::Api.configure do|c|
   c.host = 'http://dev.vle.getsmarter.co.za'
   c.token = '072556801bf07076fff6bff2a463b7c5'
 end
 
 # Set configuration values individually
-Moodle.configuration.host = 'http://dev.vle.getsmarter.co.za'
-Moodle.configuration.token = '072556801bf07076fff6bff2a463b7c5'
+Moodle::Api.configuration.host = 'http://dev.vle.getsmarter.co.za'
+Moodle::Api.configuration.token = '072556801bf07076fff6bff2a463b7c5'
 
 # Pass options hash to configure
-Moodle.configure({host: 'http://dev.vle.getsmarter.co.za',
+Moodle::Api.configure({host: 'http://dev.vle.getsmarter.co.za',
                   token: '072556801bf07076fff6bff2a463b7c5'})
 
 ```
@@ -55,7 +55,7 @@ Moodle requires users to expose [web service functions](https://docs.moodle.org/
 All web services are available by calling
 
 ```ruby
-Moodle.function_name_here(my_parameters)
+Moodle::Api.function_name_here(my_parameters)
 ```
 
 New functions created in Moodle will automatically be available in the gem.
@@ -63,26 +63,26 @@ New functions created in Moodle will automatically be available in the gem.
 ### Example
 
 ```ruby
-Moodle.configure({host: 'http://dev.vle.getsmarter.co.za',
+Moodle::Api.configure({host: 'http://dev.vle.getsmarter.co.za',
                   token: '072556801bf07076fff6bff2a463b7c5'})
 
 params = { 'criteria[0][key]' => 'firstname', 'criteria[0][value]' => 'Jon' }
-Moodle.core_user_get_users(params)
+Moodle::Api.core_user_get_users(params)
 ```
 
 ### Authentication
 Moodle uses token authentication, but sometimes you might not have a token. Users are able to generate tokens automatically when calling services using basic authentication.
 ```ruby
-Moodle.configure({host: 'http://dev.vle.getsmarter.co.za',
+Moodle::Api.configure({host: 'http://dev.vle.getsmarter.co.za',
                   service: 'my_external_service', # ensure you include the shortname of the external service
                   username: 'jonsnow',
                   password: 'defendthewall'})
 params = { 'criteria[0][key]' => 'firstname', 'criteria[0][value]' => 'Jon' }
-Moodle.core_user_get_users(params)
+Moodle::Api.core_user_get_users(params)
 ```
 The gem will handle generating the token automatically and making the call to `core_user_get_users` with the token. If you want you can also just generate a token using
 ```ruby
-configuration = Moodle::Configuration.new
+configuration = Moodle::Api::Configuration.new
 configuration.host = 'http://example.com'
 configuration.username = 'jonsnow'
 configuration.password = 'batman'
@@ -90,7 +90,7 @@ configuration.service = 'service'
 configuration
 
 # Note you could pass in a struct with these attributes and it should work the same
-Moodle::TokenGenerator.new(configuration).call
+Moodle::Api::TokenGenerator.new(configuration).call
 ```
 
 ## Documentation
