+++ lib/moodle/api/errors.rb
@@ -1,0 +1,5 @@
+module Moodle
+  module Api
+    class MoodleError < StandardError; end
+  end
+end
