+++ lib/moodle/api/version.rb
@@ -1,0 +1,5 @@
+module Moodle
+  module Api
+    VERSION = "0.1.0"
+  end
+end
