--- lib/moodle/errors.rb
@@ -1,3 +1,0 @@
-module Moodle
-  class MoodleError < StandardError; end
-end
