--- lib/moodle/version.rb
@@ -1,3 +1,0 @@
-module Moodle
-  VERSION = "0.1.0"
-end
