+++ spec/spec_helper.rb
@@ -1,4 +1,4 @@
-require 'moodle'
+require 'moodle/api'
 require 'vcr'
 
 # This file was generated by the `rspec --init` command. Conventionally, all
